Move rectangle out of cops
