+++ internal/rectangle/rectangle.go
